SaveGameFile extracted
